Add utf8 as the charset
This will tell Vesta to read data as UTF-8. This fixes the issue of attributes having to be
under a certain length. If a UTF-8 string is parsed as ASCII, then strings will go pass that
limit.

For example:
"Résidence University Val Pomp" is 30 characters in length, but actually 31 bytes.
Without UTF-8 encoding, it is read as:
"RÃ©sidence University Val Pompa"